[Refactoring] Separate AverageReview Component - Separated into AverageReview Component
